chore: reorganize content
