+++ contents/enablement-page/enablement.yml
@@ -1,0 +1,58 @@
+name: Programma di abilitazione
+title: Programma di abilitazione al cloud
+body: Strumenti, risorse e metodologie per innovare i servizi digitali pubblici utilizzando tecnologie e infrastrutture cloud.
+resourceTitleBlockOne: Kit di abilitazione al cloud (Cloud enablement kit)
+resourceTitleBlockTwo: Framework di abilitazione al cloud (Cloud enablement framework)
+resources:
+  - title: Lista e prioritizzazione degli applicativi
+    references:
+      - action: Vai alla <strong>risorsa</strong>
+        icon: it-file
+        link: https://docs.google.com/presentation/d/1ldRUvIX_1qyA94CehuQ-crQyqIj1J13oJGmJpq_Hx3c
+        ariaLabel: 'Slides - Lista e prioritizzazione degli applicativi : (Link esterno) Vai a Slides - Lista e prioritizzazione degli applicativi su Google Docs'
+        title: Slides - Lista e prioritizzazione degli applicativi
+      - action: Vai alla <strong>risorsa</strong>
+        icon: it-file
+        link: https://drive.google.com/open?id=1iVswHZ22zanOSxFoQ8lSW8B1Enx9mg9gQAPVhayerNs
+        title: Template - Lista degli applicativi
+        ariaLabel: 'Template - Lista degli applicativi: (Link esterno) Vai a Template - Lista degli applicativi su Google Docs'
+  - title: Assessment e migrazione degli applicativi
+    references:
+      - action: Vai alla <strong>risorsa</strong>
+        icon: it-file
+        link: https://docs.google.com/presentation/d/1VjHmMAug82koAGbQWoNqQepwo6UP1G2O7NzNtpx7uXo
+        ariaLabel: "Slides - Scheda di assessment dell'applicativo e scelta della strategia di migrazione : (Link esterno) Vai a Slides - Scheda di assessment dell'applicativo e scelta della strategia di migrazione su Google Docs"
+        title: Slides - Scheda di assessment dell'applicativo e scelta della strategia di migrazione
+      - action: Vai alla <strong>risorsa</strong>
+        icon: it-file
+        link: https://drive.google.com/open?id=1P8lcsCxEXKYk7oZVoHrS6rhNmqdPz4bHfzmsmWq4akQ
+        title: Template - Scheda di assessment
+        ariaLabel: "Template - Scheda di assessment : (Link esterno) Vai a Template - Scheda di assessment su Google Docs"
+  - title: Valutazione delle competenze
+    references:
+      - action: Vai alla <strong>risorsa</strong>
+        icon: it-file
+        link: https://docs.google.com/presentation/d/1De60ZAKkrBHary4iTi_FXSapWEnoG-AbAkS1AGV7EH8
+        ariaLabel: "Slides - Mappatura Competenze e Indicatori di Performance : (Link esterno) Vai a Slides - Mappatura Competenze e Indicatori di Performance su Google Docs"
+        title: Slides - Mappatura Competenze e Indicatori di Performance
+      - action: Vai alla <strong>risorsa</strong>
+        icon: it-file
+        link: https://drive.google.com/open?id=1_5aEEDg4TGDEJB8EM5ZQqb6XgGtCxISLRFa_-l2BscU
+        title: Template - Valutazione delle competenze
+        ariaLabel: "Template - Valutazione delle competenze : (Link esterno) Vai a Template - Valutazione delle competenze su Google Docs"
+resourcesFramework:
+  - references:
+      - action: Leggi su <strong>Docs Italia</strong>
+        description: Risorse, strategie operative, metodologie e strumenti necessari per attuare il programma di abilitazione al cloud della PA
+        icon: it-external-link
+        link: https://docs.italia.it/italia/piano-triennale-ict/cloud-docs/it/stabile/cloud-enablement.html#la-strategia-di-cloud-enablement
+        ariaLabel: 'Il framework di lavoro: (Link esterno) Vai a Il framework di lavoro su Docs Italia'
+        title: Il framework di lavoro
+resourcesManual:
+  - references:
+      - action: Leggi su <strong>Docs Italia</strong>
+        description: Metodi, strumenti e buone pratiche che le pubbliche amministrazioni possono usare per la migrazione al cloud di infrastrutture e applicativi esistenti
+        icon: it-external-link
+        link: https://docs.italia.it/italia/manuale-di-abilitazione-al-cloud/manuale-di-abilitazione-al-cloud-docs/it/bozza/index.html
+        ariaLabel: 'Manuale di abilitazione al cloud: (Link esterno) Vai a Manuale di abilitazione al cloud su Docs Italia'
+        title: Manuale di abilitazione al cloud
