--- contents/enablement-page/index.yml
@@ -1,5 +1,0 @@
-name: Programma di abilitazione
-heroTitle: Programma di abilitazione al cloud
-heroBody: Strumenti, risorse e metodologie per innovare i servizi digitali pubblici utilizzando tecnologie e infrastrutture cloud.
-resourceTitleBlockOne: Kit di abilitazione al cloud (Cloud enablement kit)
-resourceTitleBlockTwo: Framework di abilitazione al cloud (Cloud enablement framework)
