--- contents/enablement-page/resources.yml
@@ -1,36 +1,0 @@
-- title: Lista e prioritizzazione degli applicativi
-  references:
-    - action: Vai alla <strong>risorsa</strong>
-      icon: it-file
-      link: https://docs.google.com/presentation/d/1ldRUvIX_1qyA94CehuQ-crQyqIj1J13oJGmJpq_Hx3c
-      ariaLabel: 'Slides - Lista e prioritizzazione degli applicativi : (Link esterno) Vai a Slides - Lista e prioritizzazione degli applicativi su Google Docs'
-      title: Slides - Lista e prioritizzazione degli applicativi
-    - action: Vai alla <strong>risorsa</strong>
-      icon: it-file
-      link: https://drive.google.com/open?id=1iVswHZ22zanOSxFoQ8lSW8B1Enx9mg9gQAPVhayerNs
-      title: Template - Lista degli applicativi
-      ariaLabel: 'Template - Lista degli applicativi: (Link esterno) Vai a Template - Lista degli applicativi su Google Docs'
-- title: Assessment e migrazione degli applicativi
-  references:
-    - action: Vai alla <strong>risorsa</strong>
-      icon: it-file
-      link: https://docs.google.com/presentation/d/1VjHmMAug82koAGbQWoNqQepwo6UP1G2O7NzNtpx7uXo
-      ariaLabel: "Slides - Scheda di assessment dell'applicativo e scelta della strategia di migrazione : (Link esterno) Vai a Slides - Scheda di assessment dell'applicativo e scelta della strategia di migrazione su Google Docs"
-      title: Slides - Scheda di assessment dell'applicativo e scelta della strategia di migrazione
-    - action: Vai alla <strong>risorsa</strong>
-      icon: it-file
-      link: https://drive.google.com/open?id=1P8lcsCxEXKYk7oZVoHrS6rhNmqdPz4bHfzmsmWq4akQ
-      title: Template - Scheda di assessment
-      ariaLabel: "Template - Scheda di assessment : (Link esterno) Vai a Template - Scheda di assessment su Google Docs"
-- title: Valutazione delle competenze
-  references:
-    - action: Vai alla <strong>risorsa</strong>
-      icon: it-file
-      link: https://docs.google.com/presentation/d/1De60ZAKkrBHary4iTi_FXSapWEnoG-AbAkS1AGV7EH8
-      ariaLabel: "Slides - Mappatura Competenze e Indicatori di Performance : (Link esterno) Vai a Slides - Mappatura Competenze e Indicatori di Performance su Google Docs"
-      title: Slides - Mappatura Competenze e Indicatori di Performance
-    - action: Vai alla <strong>risorsa</strong>
-      icon: it-file
-      link: https://drive.google.com/open?id=1_5aEEDg4TGDEJB8EM5ZQqb6XgGtCxISLRFa_-l2BscU
-      title: Template - Valutazione delle competenze
-      ariaLabel: "Template - Valutazione delle competenze : (Link esterno) Vai a Template - Valutazione delle competenze su Google Docs"
