--- contents/home-page/benefits.yml
@@ -1,13 +1,0 @@
-- title: Per i cittadini
-  body: Servizi pubblici basati sul modello cloud garantiscono ai cittadini maggiore affidabilità, sicurezza e rispetto
-    della privacy. Sono servizi progettati in maniera nativa digitale e che hanno una minore incidenza sulla spesa
-    pubblica.
-- title: Per le amministrazioni
-  body: Le amministrazioni beneficiano di risparmi significativi da reinvestire nello sviluppo di nuovi servizi,
-    maggiore trasparenza sui costi e sull’utilizzo dei servizi, agilità e scalabilità nella gestione delle
-    infrastrutture.
-- title: Per le PMI dell’ICT che operano nel settore pubblico
-  body: Le piccole e medie imprese (PMI) nel settore delle tecnologie digitali sono chiamate ad accompagnare le
-    amministrazioni verso l’adozione di soluzioni in cloud per i propri servizi, aumentando la qualità e la quantità
-    di servizi qualificati. Rappresentano un elemento di stimolo per il settore pubblico nel rivolgersi a un mercato
-    maturo per le soluzioni cloud.
