--- contents/home-page/contribute.yml
@@ -1,15 +1,0 @@
-- ariaLabel: 'Forum Italia: (Link esterno) Vai Forum Italia'
-  body: Entra in <strong>forum.italia.it</strong> e condividi le tue opinioni nella sezione <strong>Cloud e data center</strong>
-  icon: it-designers-italia
-  link: https://forum.italia.it/
-  title: Forum
-- ariaLabel: 'Slack Developers Italia: (Link esterno) Vai Slack Developers Italia'
-  body: Unisciti alla community di <strong>Developers Italia su Slack</strong> e raggiungi il canale <strong>#cloud</strong>
-  icon: it-star-full
-  link: https://slack.developers.italia.it/
-  title: Slack
-- ariaLabel: 'GitHub Italia: (Link esterno) Vai GitHub Italia'
-  body: Proponi <em>issue</em>, modifiche e nuovi contenuti nei <strong>repository pubblici</strong> che contengono codice e documentazione per lo sviluppo del cloud della PA
-  icon: it-github
-  link: https://github.com/italia
-  title: Github
