--- contents/home-page/news-preview.yml
@@ -1,18 +1,0 @@
-- body: Avviato il tavolo di lavoro tra il Dipartimento e gli Enti centrali dello Stato proprietari di infrastrutture digitali.
-  date: 27 novembre 2020
-  link: https://innovazione.gov.it/notizie/articoli/sviluppiamo-il-sistema-operativo-a-partire-dalle-infrastrutture-digitali/
-  source: 'Fonte: innovazione.gov.it'
-  title: Sviluppiamo il Sistema Operativo a partire dalle Infrastrutture Digitali
-  type: Articolo
-- body: Avviato il tavolo di lavoro tra il Dipartimento e gli Enti centrali dello Stato proprietari di infrastrutture digitali.
-  date: 21 febbraio 2020
-  link: https://medium.com/blog-per-la-trasformazione-digitale/inizia-la-rivoluzione-cloud-la-strategia-per-le-infrastrutture-digitali-della-pubblica-eddcd0a8e3f8
-  source: 'Fonte: medium'
-  title: 'Inizia la rivoluzione cloud: la strategia per le infrastrutture digitali della Pubblica Amministrazione'
-  type: Approfondimento
-- body: Avviato il tavolo di lavoro tra il Dipartimento e gli Enti centrali dello Stato proprietari di infrastrutture digitali.
-  date: 12 luglio 2019
-  link: https://medium.com/team-per-la-trasformazione-digitale/migrazione-in-cloud-servizi-pubblica-amministrazione-vantaggi-rischi-strategie-44ec2439bebf
-  source: 'Fonte: medium'
-  title: Un kit per portare i servizi pubblici nel cloud
-  type: Approfondimento
