--- contents/strategy-page/index.yml
@@ -1,8 +1,0 @@
-name: Strategia cloud
-heroTitle: 'La strategia nazionale cloud della PA'
-heroBody: 'Il Piano triennale per l’informatica nella pubblica amministrazione prevede la realizzazione del sistema
-          operativo del paese mediante l’adozione del modello cloud computing nelle pubbliche amministrazioni. Il
-          Dipartimento per la trasformazione digitale, in collaborazione con AgID, ha definito la strategia Cloud
-          della PA e il programma per abilitare il cloud computing nel settore pubblico.'
-timelineTitle: 'Piano di attuazione'
-resourceTitle: 'Risorse utili'
