+++ gatsby-ssr.js
@@ -1,5 +1,5 @@
 import React from 'react';
-import { Layout } from './src/components/Layout.js';
+import { Layout } from './src/layouts/Layout.js';
 
 export const wrapPageElement = ({ element, props }) => (
   <Layout {...props}>{element}</Layout>
