+++ package.json
@@ -39,7 +39,6 @@
     "gatsby-source-filesystem": "^3.0.0",
     "gatsby-transformer-remark": "^3.0.0",
     "gatsby-transformer-sharp": "^3.0.0",
-    "gatsby-transformer-yaml": "^3.0.0",
     "gh-pages": "^3.1.0",
     "remark-attr": "^0.11.1"
   }
