+++ src/layouts/EnablementPage.js
@@ -1,0 +1,92 @@
+import React from 'react';
+import { StaticImage } from 'gatsby-plugin-image';
+import { createUseStyles } from 'react-jss';
+import { graphql, useStaticQuery } from 'gatsby';
+import { HeroTitle } from '../components/hero/HeroTitle.js';
+import { HeroBody } from '../components/hero/HeroBody.js';
+import { Hero } from '../components/hero/Hero.js';
+import { ResourcesWithAccordion } from '../components/resource/ResourcesWithAccordion.js';
+import { ResourcesWithList } from '../components/resource/ResourcesWithList.js';
+import { ResourceTitle } from '../components/resource/ResourceTitle.js';
+import { TextChunk } from '../components/TextChunk.js';
+import content from '../../contents/enablement-page/enablement.yml';
+
+const useStyles = createUseStyles({
+  noBorderBottom: {
+    '@global': {
+      '.it-right-zone': {
+        borderBottom: [['none'], '!important'],
+      },
+    },
+  },
+});
+
+const {
+  name,
+  title,
+  body,
+  resourceTitleBlockOne,
+  resourceTitleBlockTwo,
+  resources,
+  resourcesManual,
+  resourcesFramework,
+} = content;
+
+const query = graphql`
+  query {
+    textChunk: markdownRemark(fields: { slug: { eq: "enablement-page/enablement.md" } }) {
+      html
+    }
+  }
+`;
+
+const EnablementPage = () => {
+  const classes = useStyles();
+  const {
+    textChunk: { html: textChunk },
+  } = useStaticQuery(query);
+
+  return (
+    <>
+      <h1 className="sr-only">{name}</h1>
+      <Hero>
+        <div className="row align-items-center">
+          <div className="offset-lg-1 col-lg-6 mt-4 mt-lg-0">
+            <div className="text-center text-lg-left">
+              <HeroTitle title={title} className="text-info" Tag="h2" />
+              <HeroBody html={body} />
+            </div>
+          </div>
+        </div>
+      </Hero>
+
+      <StaticImage
+        aspectRatio={1280 / 548}
+        src="../images/strategia_hero_cloud_2x.jpg"
+        alt=""
+        placeholder="blurred"
+        formats={['AUTO', 'AVIF', 'WEBP']}
+      />
+
+      <Hero>
+        <TextChunk html={textChunk} />
+      </Hero>
+
+      <Hero bgColor="light">
+        <a name="kit" />
+        <div className="col-xl-8 col-lg-10 mx-auto">
+          <ResourceTitle title={resourceTitleBlockOne} />
+          <ResourcesWithList className={`${classes.noBorderBottom} mt-4`} resources={resourcesManual} />
+          <ResourcesWithAccordion resources={resources} />
+        </div>
+        <a name="framework" />
+        <div className="col-xl-8 col-lg-10 mx-auto mt-5">
+          <ResourceTitle title={resourceTitleBlockTwo} />
+          <ResourcesWithList className={`${classes.noBorderBottom} mt-4`} resources={resourcesFramework} />
+        </div>
+      </Hero>
+    </>
+  );
+};
+
+export default EnablementPage;
