+++ src/layouts/Footer.js
@@ -4,7 +4,7 @@ import { createUseStyles } from 'react-jss';
 import dtdLogo from '../images/dtd-logo.svg';
 import agidLogo from '../images/agid-logo.svg';
 import links from '../../contents/links.yml';
-import { ExternalLink } from './ExternalLink.js';
+import { ExternalLink } from '../components/ExternalLink.js';
 
 const {
   internalLinks: { glossary },
