+++ src/layouts/home/HowToContributeSection.js
@@ -1,17 +1,17 @@
 import React from 'react';
 import { Card, CardBody, Icon } from 'design-react-kit';
-import { MobileSwiper } from '../MobileSwiper.js';
-import { Hero } from '../hero/Hero.js';
-import { ExternalLink } from '../ExternalLink.js';
-import heroContent from '../../../contents/home-page/index.yml';
-import contributeCards from '../../../contents/home-page/contribute.yml';
+import { MobileSwiper } from '../../components/MobileSwiper.js';
+import { Hero } from '../../components/hero/Hero.js';
+import { ExternalLink } from '../../components/ExternalLink.js';
+import content from '../../../contents/home-page/index.yml';
 
 const {
   heroContribute: { category, title, body },
-} = heroContent;
+  contributions,
+} = content;
 
-export const HowToContribute = () => {
-  const slides = contributeCards.map((card) => (
+export const HowToContributeSection = () => {
+  const slides = contributions.map((card) => (
     <Card key={card.title} teaser noWrapper className="rounded shadow-lg col-lg-3 col-12 mr-4">
       <ExternalLink linkTo={card.link} ariaLabel={card.ariaLabel} className="text-decoration-none">
         <CardBody className="pb-5">
